@@ -16,10 +16,10 @@
     "build-storybook": "build-storybook"
   },
   "dependencies": {
-    "@babel/plugin-transform-runtime": "~7.16.8",
+    "@babel/plugin-transform-runtime": "~7.17.0",
     "@sentry/browser": "~6.17.3",
-    "@visx/axis": "~2.4.0",
-    "@visx/event": "~2.1.2",
+    "@visx/axis": "~2.6.0",
+    "@visx/event": "~2.6.0",
     "@visx/glyph": "~2.1.2",
     "@visx/group": "~2.1.0",
     "@visx/responsive": "~2.8.0",
@@ -31,7 +31,7 @@
     "cuid": "~2.1.6",
     "d3": "~6.7.0",
     "graphql": "~16.3.0",
-    "graphql-request": "~3.7.0",
+    "graphql-request": "~4.0.0",
     "hash.js": "~1.1.7",
     "i18next": "~21.6.3",
     "lodash": "~4.17.11",
@@ -54,13 +54,13 @@
     "styled-components": "~5.x.x"
   },
   "devDependencies": {
-    "@babel/core": "~7.16.7",
-    "@babel/plugin-proposal-decorators": "~7.16.7",
+    "@babel/core": "~7.17.0",
+    "@babel/plugin-proposal-decorators": "~7.17.0",
     "@babel/plugin-proposal-logical-assignment-operators": "~7.16.7",
     "@babel/polyfill": "~7.12.1",
     "@babel/preset-env": "~7.16.8",
     "@babel/preset-react": "~7.16.7",
-    "@babel/register": "~7.16.9",
+    "@babel/register": "~7.17.0",
     "@storybook/addon-a11y": "~6.4.9",
     "@storybook/addon-essentials": "~6.4.9",
     "@storybook/addon-knobs": "~6.4.0",
@@ -83,7 +83,7 @@
     "babel-plugin-webpack-alias": "~2.1.2",
     "chai": "~4.3.4",
     "chai-dom": "~1.11.0",
-    "css-loader": "~6.5.1",
+    "css-loader": "~6.6.0",
     "dirty-chai": "~2.0.1",
     "enzyme": "~3.11.0",
     "grommet": "~2.17.4",
@@ -105,14 +105,14 @@
     "react": "~17.0.2",
     "react-dom": "~17.0.2",
     "rosie": "~2.1.0",
-    "sinon": "~12.0.1",
+    "sinon": "~13.0.1",
     "sinon-chai": "~3.7.0",
     "snazzy": "~9.0.0",
     "storybook-react-i18next": "~1.0.16",
     "style-loader": "~3.3.1",
     "styled-components": "~5.3.3",
-    "superagent": "~5.1.0",
-    "webpack": "~5.66.0",
+    "superagent": "~7.1.1",
+    "webpack": "~5.68.0",
     "webpack-cli": "~4.9.1",
     "webpack-dev-server": "~4.7.3"
   },
