@@ -47,10 +47,10 @@
     "styled-components": "5.x.x"
   },
   "devDependencies": {
-    "@babel/core": "~7.16.7",
+    "@babel/core": "~7.17.0",
     "@babel/preset-env": "~7.16.8",
     "@babel/preset-react": "~7.16.7",
-    "@babel/register": "~7.16.9",
+    "@babel/register": "~7.17.0",
     "@storybook/addon-a11y": "~6.4.9",
     "@storybook/addon-essentials": "~6.4.9",
     "@storybook/addon-knobs": "~6.4.0",
@@ -69,7 +69,7 @@
     "chai": "~4.3.4",
     "chai-dom": "~1.11.0",
     "clean-webpack-plugin": "~4.0.0",
-    "css-loader": "~6.5.1",
+    "css-loader": "~6.6.0",
     "dedent": "~0.7.0",
     "dirty-chai": "~2.0.1",
     "enzyme": "~3.11.0",
@@ -83,12 +83,12 @@
     "polished": "~4.1.4",
     "react": "~17.0.2",
     "react-dom": "~17.0.2",
-    "sinon": "~12.0.1",
+    "sinon": "~13.0.1",
     "sinon-chai": "~3.7.0",
     "snazzy": "~9.0.0",
     "style-loader": "~3.3.1",
     "styled-components": "~5.3.3",
-    "webpack": "~5.66.0",
+    "webpack": "~5.68.0",
     "webpack-cli": "~4.9.1"
   },
   "engines": {
