@@ -1,8 +1,14 @@
 import getDefaultPageProps from '@helpers/getDefaultPageProps'
 export { default } from '@screens/ProjectHomePage'
+import { serverSideTranslations } from 'next-i18next/serverSideTranslations'
 
-export async function getServerSideProps({ params, query, req, res }) {
+export async function getServerSideProps({ locale, params, query, req, res }) {
   const { notFound, props } = await getDefaultPageProps({ params, query, req, res })
-  return ({ notFound, props })
+  return ({ 
+    notFound,
+    props: {
+      ...(await serverSideTranslations(locale, ['screens'])),
+      ...props
+    }
+  })
 }
-
