@@ -1,14 +1,11 @@
-import counterpart from 'counterpart'
 import PropTypes from 'prop-types'
 import { withTheme } from 'styled-components'
 import { Media, ZooniverseLogo } from '@zooniverse/react-components'
-
-import en from './locales/en'
-
-counterpart.registerTranslations('en', en)
+import { useTranslation } from 'next-i18next'
 
 function ProjectImage ({ imageSrc, projectName, theme }) {
-  const alt = counterpart('ProjectImage.alt', { projectName })
+  const { t } = useTranslation('screens')
+  const alt = t('Classify.FinishedForTheDay.ProjectImage.alt', { projectName })
   return (
     <Media
       alt={alt}
