@@ -2,14 +2,12 @@ import PropTypes from 'prop-types'
 import { Button, Box, CheckBox, Text } from 'grommet'
 import { Modal, PrimaryButton, SpacedText } from '@zooniverse/react-components'
 import { useRouter } from 'next/router'
-import counterpart from 'counterpart'
+import { useTranslation } from 'next-i18next'
 
 import NavLink from '@shared/components/NavLink'
-import en from './locales/en'
-
-counterpart.registerTranslations('en', en)
 
 export default function WorkflowAssignmentModal(props) {
+  const { t } = useTranslation('screens')
   const { active = false, assignedWorkflowID, closeFn, dismiss, dismissedForSession = false } = props
   const router = useRouter()
   const { owner, project } = router?.query || {}
@@ -17,21 +15,21 @@ export default function WorkflowAssignmentModal(props) {
   const url = `/${owner}/${project}/classify/workflow/${assignedWorkflowID}`
 
   return (
-    <Modal active={active} closeFn={closeFn} title={counterpart('WorkflowAssignmentModal.title')}>
-      <Box pad={{ bottom: 'xsmall' }}>{counterpart('WorkflowAssignmentModal.content')}</Box>
+    <Modal active={active} closeFn={closeFn} title={t('Classify.WorkflowAssignmentModal.title')}>
+      <Box pad={{ bottom: 'xsmall' }}>{t('Classify.WorkflowAssignmentModal.content')}</Box>
       <Box pad={{ bottom: 'xsmall' }}>
         <CheckBox
           checked={dismissedForSession}
-          label={<SpacedText>{counterpart('WorkflowAssignmentModal.dismiss')}</SpacedText>}
+          label={<SpacedText>{t('Classify.WorkflowAssignmentModal.dismiss')}</SpacedText>}
           onChange={(event) => dismiss(event)}
         />
       </Box>
       <Box direction='row' gap='xsmall' justify='center'>
-        <Button label={counterpart('WorkflowAssignmentModal.cancel')} onClick={() => closeFn()} />
+        <Button label={t('Classify.WorkflowAssignmentModal.cancel')} onClick={() => closeFn()} />
         <NavLink
           StyledAnchor={PrimaryButton}
           StyledSpacedText={Text}
-          link={{ href: url, text: counterpart('WorkflowAssignmentModal.confirm')}}
+          link={{ href: url, text: t('Classify.WorkflowAssignmentModal.confirm') }}
         />
       </Box>
     </Modal>
