@@ -5,6 +5,14 @@ import RecentSubjectsContainer from './RecentSubjectsContainer'
 
 describe('Component > RecentSubjectsContainer', function () {
   let wrapper
+  const stores = {
+    store: {
+      project: {
+        id: '1234',
+        slug: '/test-owner/test-project'
+      }
+    }
+  }
 
   const TALK_URL = 'https://talk-staging.zooniverse.org'
   const PANOPTES_URL = 'https://panoptes-staging.zooniverse.org/api'
@@ -31,7 +39,7 @@ describe('Component > RecentSubjectsContainer', function () {
       .get('/subjects')
       .query(true)
       .reply(200, { subjects: MOCK_SUBJECTS })
-    wrapper = shallow(<RecentSubjectsContainer.wrappedComponent />)
+    wrapper = shallow(<RecentSubjectsContainer stores={stores} />)
   })
 
   after(function () {
