@@ -71,6 +71,20 @@ function useStore({ authClient, client, initialState }) {
   return _store
 }
 
+async function fetchWorkflow(workflowID) {
+  if (workflowID) {
+    const { body } = await panoptesClient.get(`/workflows/${workflowID}`)
+    const [ workflowSnapshot ] = body.workflows
+    return workflowSnapshot
+  }
+  return null
+}
+
+function useWorkflowSnapshot(workflowID) {
+  const { data } = useSWR(workflowID, fetchWorkflow)
+  return data ?? null
+}
+
 export default function ClassifierContainer({
   authClient,
   cachePanoptesData = false,
@@ -94,11 +108,7 @@ export default function ClassifierContainer({
   })
 
   const [loaded, setLoaded] = useState(false)
-  const { data } = useSWR(`/workflows/${workflowID}`, client.panoptes.get)
-  let workflowData
-  if (data?.text) {
-    workflowData = data.text
-  }
+  const workflowSnapshot = useWorkflowSnapshot(workflowID)
 
   async function onMount() {
     if (cachePanoptesData) {
@@ -149,24 +159,26 @@ export default function ClassifierContainer({
   }, [loaded, authClient])
 
   try {
-    if (!loaded) {
-      return <Paragraph>Loading…</Paragraph>
+    if (loaded) {
+
+      return (
+        <Provider classifierStore={classifierStore}>
+          <Classifier
+            classifierStore={classifierStore}
+            locale={locale}
+            onError={onError}
+            project={project}
+            subjectSetID={subjectSetID}
+            subjectID={subjectID}
+            workflowSnapshot={workflowSnapshot}
+            workflowVersion={workflowSnapshot?.version}
+            workflowID={workflowSnapshot?.id}
+          />
+        </Provider>
+      )
     }
 
-    return (
-      <Provider classifierStore={classifierStore}>
-        <Classifier
-          classifierStore={classifierStore}
-          locale={locale}
-          onError={onError}
-          project={project}
-          subjectSetID={subjectSetID}
-          subjectID={subjectID}
-          workflowData={workflowData}
-          workflowID={workflowID}
-        />
-      </Provider>
-    )
+    return <Paragraph>Loading…</Paragraph>
   } catch (error) {
     const info = {
       package: '@zooniverse/classifier'
