@@ -1,18 +1,16 @@
-import counterpart from 'counterpart'
 import PropTypes from 'prop-types'
 import React from 'react'
+import { useTranslation } from 'react-i18next'
 
 import PointerIcon from './PointerIcon'
-import en from './locales/en'
 import Button from '../Button'
 
-counterpart.registerTranslations('en', en)
-
 function AnnotateButton ({ active, onClick }) {
+  const { t } = useTranslation('components')
   return (
     <Button
       active={active}
-      a11yTitle={counterpart('AnnotateButton.ariaLabel')}
+      a11yTitle={t('ImageToolbar.AnnotateButton.ariaLabel')}
       icon={<PointerIcon />}
       onClick={onClick}
     />
@@ -26,7 +24,7 @@ AnnotateButton.propTypes = {
 
 AnnotateButton.defaultProps = {
   active: false,
-  onClick: () => console.log(counterpart('AnnotateButton.ariaLabel'))
+  onClick: () => console.log('Annotate Button')
 }
 
 export default AnnotateButton
