@@ -1,18 +1,16 @@
-import counterpart from 'counterpart'
 import PropTypes from 'prop-types'
 import React from 'react'
+import { useTranslation } from 'react-i18next'
 
 import ActualSizeIcon from './ActualSizeIcon'
 import FullscreenIcon from './FullscreenIcon'
-import en from './locales/en'
 import Button from '../Button'
 
-counterpart.registerTranslations('en', en)
-
 function FullscreenButton ({ active, disabled, onClick }) {
+  const { t } = useTranslation('components')
   const Icon = (active) ? ActualSizeIcon : FullscreenIcon
   const labelKey = active ? 'actualSize' : 'fullscreen'
-  const label = counterpart(`FullscreenButton.ariaLabel.${labelKey}`)
+  const label = t(`ImageToolbar.FullscreenButton.ariaLabel.${labelKey}`)
 
   return (
     <Button
