@@ -1,17 +1,15 @@
-import counterpart from 'counterpart'
 import PropTypes from 'prop-types'
 import React from 'react'
+import { useTranslation } from 'react-i18next'
 
 import MoveIcon from './MoveIcon'
-import en from './locales/en'
 import Button from '../Button'
 
-counterpart.registerTranslations('en', en)
-
 function MoveButton ({ active, onClick }) {
+  const { t } = useTranslation('components')
   return (
     <Button
-      a11yTitle={counterpart('MoveButton.ariaLabel')}
+      a11yTitle={t('ImageToolbar.MoveButton.ariaLabel')}
       active={active}
       icon={<MoveIcon />}
       onClick={onClick}
@@ -26,7 +24,7 @@ MoveButton.propTypes = {
 
 MoveButton.defaultProps = {
   active: false,
-  onClick: () => console.log(counterpart('MoveButton.ariaLabel'))
+  onClick: () => console.log('Move Button')
 }
 
 export default MoveButton
