@@ -2,6 +2,7 @@ import { shallow } from 'enzyme'
 import React from 'react'
 import sinon from 'sinon'
 import ZoomInButton from './ZoomInButton'
+import i18n from '@test/i18n/i18n-for-tests'
 
 describe('Component > ZoomInButton', function () {
   it('should render without crashing', function () {
@@ -10,8 +11,11 @@ describe('Component > ZoomInButton', function () {
   })
 
   it('should have an `a11yTitle` label', function () {
+    const useTranslationStub = sinon.stub(i18n, 't').callsFake((key) => key)
     const wrapper = shallow(<ZoomInButton />)
-    expect(wrapper.prop('a11yTitle')).to.equal('Zoom in on subject')
+    expect(wrapper.prop('a11yTitle')).exists()
+    expect(useTranslationStub).to.have.been.calledWith('ImageToolbar.ZoomInButton.ariaLabel')
+    useTranslationStub.restore()
   })
 
   it('should call the onClick prop function on click', function () {
